Some one please check my changes - numbers are too high
We got more FFF WORK

+++ src/letter.java
@@ -1,6 +1,6 @@
 //The data collected to a single letter
 public class letter {
-	public float avgPressTime;
+	public int avgPressTime;
 	private int timesPressed;
 
 	letter() {
